Finished return & Cancel Functions. Fixed Imperfections.
Fixed the return functionality by having "TaxedReturnTotal" variable update every time a new product is scanned, effectively allowing you to hit return and still add items to the returns without having to hit the return button again. Also, the cancel function would only delete the sum of the prices * quantity of the products with the status of Status = "Sale" but it now accounts for returned products prices as well. Lastly, set "Cash_Number_Label" to "0.00" every time a discount is added or a product is scanned.